+++ src/elements/JobsPanelVolumesTable/JobsPanelVolumesTable.js
@@ -10,8 +10,8 @@ import JobsPanelTableAddItemRow from '../JobsPanelTableAddItemRow/JobsPanelTable
 import JobsPanelTable from '../JobsPanelTable/JobsPanelTable'
 
 import panelData from '../../components/JobsPanel/panelData'
-import { inputsActions } from '../../components/JobsPanelDataInputs/inputsReducer'
-import { selectOptions } from '../../components/JobsPanelDataInputs/inputs.util'
+import { inputsActions } from '../../components/JobsPanelDataInputs/jobsPanelDataInputsReducer'
+import { selectOptions } from '../../components/JobsPanelDataInputs/jobsPanelDataInputs.util'
 
 import { ReactComponent as Plus } from '../../images/plus.svg'
